Refactor: Extract duplicated `TableWriter` into `utils.py`; use function to extract duplicated code for generating options table.
