+++ docs/exts/generate_message_list.py
@@ -1,38 +1,11 @@
 """Helper script to generate a table of all messages that gherlint emits."""
 
-from typing import Dict, List
-
 from sphinx.application import Sphinx
+from utils import TableWriter
 
 from gherlint.checkers.base_checker import BaseChecker
 from gherlint.linter import GherkinLinter
 from gherlint.registry import CheckerRegistry
-
-
-class TableWriter:
-    def __init__(self) -> None:
-        self._lines: List[str] = []
-
-    def add_directive(self) -> None:
-        self._lines.append(".. list-table::")
-        self._lines.append("   :header-rows: 1")
-        self._lines.append("")
-
-    def add_header(self, *headings: str) -> None:
-        self.add_row(*headings)
-
-    def add_row(self, *values: str) -> None:
-        first = True
-        for value in values:
-            if first:
-                content = f"   * - {value}"
-                first = False
-            else:
-                content = f"     - {value}"
-            self._lines.append(content)
-
-    def __str__(self) -> str:
-        return "\n".join(self._lines)
 
 
 def main(app: Sphinx) -> None:
@@ -59,6 +32,6 @@ def _get_msg_prefix(checker: BaseChecker) -> int:
     return int(checker.MESSAGES[0].id[1:-2])
 
 
-def setup(app: Sphinx) -> Dict[str, str]:
+def setup(app: Sphinx) -> dict[str, str]:
     app.connect("builder-inited", main)
     return {"version": "0.1"}
